# ExecutionResult wrapping can be avoided - reduces memory allocations and GC