--- a/src/main/java/graphql/execution/instrumentation/ChainedInstrumentation.java
+++ b/src/main/java/graphql/execution/instrumentation/ChainedInstrumentation.java
@@ -176,6 +176,21 @@ public class ChainedInstrumentation implements Instrumentation {
     }
 
     @Override
+    public @Nullable ExecuteObjectInstrumentationContext beginExecuteObject(InstrumentationExecutionStrategyParameters parameters, InstrumentationState state) {
+        if (instrumentations.isEmpty()) {
+            return ExecuteObjectInstrumentationContext.NOOP;
+        }
+        Function<Instrumentation, ExecuteObjectInstrumentationContext> mapper = instrumentation -> {
+            InstrumentationState specificState = getSpecificState(instrumentation, state);
+            return instrumentation.beginExecuteObject(parameters, specificState);
+        };
+        if (instrumentations.size() == 1) {
+            return mapper.apply(instrumentations.get(0));
+        }
+        return new ChainedExecuteObjectInstrumentationContext(mapAndDropNulls(instrumentations, mapper));
+    }
+
+    @Override
     @NotNull
     public InstrumentationContext<ExecutionResult> beginSubscribedFieldEvent(InstrumentationFieldParameters parameters) {
         return Assert.assertShouldNeverHappen("The deprecated " + "beginSubscribedFieldEvent" + " was called");
@@ -197,9 +212,14 @@ public class ChainedInstrumentation implements Instrumentation {
 
     @Override
     public InstrumentationContext<ExecutionResult> beginField(InstrumentationFieldParameters parameters, InstrumentationState state) {
+        return Assert.assertShouldNeverHappen("The deprecated " + "beginField" + " was called");
+    }
+
+    @Override
+    public @Nullable InstrumentationContext<Object> beginFieldExecution(InstrumentationFieldParameters parameters, InstrumentationState state) {
         return chainedCtx(instrumentation -> {
             InstrumentationState specificState = getSpecificState(instrumentation, state);
-            return instrumentation.beginField(parameters, specificState);
+            return instrumentation.beginFieldExecution(parameters, specificState);
         });
     }
 
@@ -226,11 +246,17 @@ public class ChainedInstrumentation implements Instrumentation {
 
     @Override
     public InstrumentationContext<ExecutionResult> beginFieldComplete(InstrumentationFieldCompleteParameters parameters, InstrumentationState state) {
+        return Assert.assertShouldNeverHappen("The deprecated " + "beginFieldComplete" + " was called");
+    }
+
+    @Override
+    public @Nullable InstrumentationContext<Object> beginFieldCompletion(InstrumentationFieldCompleteParameters parameters, InstrumentationState state) {
         return chainedCtx(instrumentation -> {
             InstrumentationState specificState = getSpecificState(instrumentation, state);
-            return instrumentation.beginFieldComplete(parameters, specificState);
+            return instrumentation.beginFieldCompletion(parameters, specificState);
         });
     }
+
 
     @Override
     @NotNull
@@ -243,6 +269,14 @@ public class ChainedInstrumentation implements Instrumentation {
         return chainedCtx(instrumentation -> {
             InstrumentationState specificState = getSpecificState(instrumentation, state);
             return instrumentation.beginFieldListComplete(parameters, specificState);
+        });
+    }
+
+    @Override
+    public @Nullable InstrumentationContext<Object> beginFieldListCompletion(InstrumentationFieldCompleteParameters parameters, InstrumentationState state) {
+        return chainedCtx(instrumentation -> {
+            InstrumentationState specificState = getSpecificState(instrumentation, state);
+            return instrumentation.beginFieldListCompletion(parameters, specificState);
         });
     }
 
@@ -371,10 +405,6 @@ public class ChainedInstrumentation implements Instrumentation {
             }
         }
 
-        private InstrumentationState getState(Instrumentation instrumentation) {
-            return instrumentationToStates.get(instrumentation);
-        }
-
         private static CompletableFuture<InstrumentationState> combineAll(List<Instrumentation> instrumentations, InstrumentationCreateStateParameters parameters) {
             Async.CombinedBuilder<InstrumentationState> builder = Async.ofExpectedSize(instrumentations.size());
             for (Instrumentation instrumentation : instrumentations) {
@@ -383,6 +413,10 @@ public class ChainedInstrumentation implements Instrumentation {
                 builder.add(stateCF);
             }
             return builder.await().thenApply(instrumentationStates -> new ChainedInstrumentationState(instrumentations, instrumentationStates));
+        }
+
+        private InstrumentationState getState(Instrumentation instrumentation) {
+            return instrumentationToStates.get(instrumentation);
         }
     }
 
@@ -434,5 +468,33 @@ public class ChainedInstrumentation implements Instrumentation {
         }
     }
 
+    private static class ChainedExecuteObjectInstrumentationContext implements ExecuteObjectInstrumentationContext {
+
+        private final ImmutableList<ExecuteObjectInstrumentationContext> contexts;
+
+        ChainedExecuteObjectInstrumentationContext(ImmutableList<ExecuteObjectInstrumentationContext> contexts) {
+            this.contexts = contexts;
+        }
+
+        @Override
+        public void onDispatched(CompletableFuture<Map<String, Object>> result) {
+            contexts.forEach(context -> context.onDispatched(result));
+        }
+
+        @Override
+        public void onCompleted(Map<String, Object> result, Throwable t) {
+            contexts.forEach(context -> context.onCompleted(result, t));
+        }
+
+        @Override
+        public void onFieldValuesInfo(List<FieldValueInfo> fieldValueInfoList) {
+            contexts.forEach(context -> context.onFieldValuesInfo(fieldValueInfoList));
+        }
+
+        @Override
+        public void onFieldValuesException() {
+            contexts.forEach(ExecuteObjectInstrumentationContext::onFieldValuesException);
+        }
+    }
 }