--- a/src/main/java/graphql/execution/instrumentation/Instrumentation.java
+++ b/src/main/java/graphql/execution/instrumentation/Instrumentation.java
@@ -5,6 +5,8 @@ import graphql.ExecutionInput;
 import graphql.ExecutionResult;
 import graphql.PublicSpi;
 import graphql.execution.ExecutionContext;
+import graphql.execution.instrumentation.adapters.ExecutionResultInstrumentationContextAdapter;
+import graphql.execution.instrumentation.adapters.ExecuteObjectInstrumentationContextAdapter;
 import graphql.execution.instrumentation.parameters.InstrumentationCreateStateParameters;
 import graphql.execution.instrumentation.parameters.InstrumentationExecuteOperationParameters;
 import graphql.execution.instrumentation.parameters.InstrumentationExecutionParameters;
@@ -201,7 +203,7 @@ public interface Instrumentation {
 
     /**
      * This is called each time an {@link graphql.execution.ExecutionStrategy} is invoked, which may be multiple times
-     * per query as the engine recursively descends down over the query.
+     * per query as the engine recursively descends over the query.
      *
      * @param parameters the parameters to this step
      *
@@ -218,7 +220,7 @@ public interface Instrumentation {
 
     /**
      * This is called each time an {@link graphql.execution.ExecutionStrategy} is invoked, which may be multiple times
-     * per query as the engine recursively descends down over the query.
+     * per query as the engine recursively descends over the query.
      *
      * @param parameters the parameters to this step
      * @param state      the state created during the call to {@link #createState(InstrumentationCreateStateParameters)}
@@ -228,6 +230,20 @@ public interface Instrumentation {
     @Nullable
     default ExecutionStrategyInstrumentationContext beginExecutionStrategy(InstrumentationExecutionStrategyParameters parameters, InstrumentationState state) {
         return beginExecutionStrategy(parameters.withNewState(state));
+    }
+
+    /**
+     * This is called each time an {@link graphql.execution.ExecutionStrategy} object resolution is called, which may be multiple times
+     * per query as the engine recursively descends over the query.
+     *
+     * @param parameters the parameters to this step
+     * @param state      the state created during the call to {@link #createState(InstrumentationCreateStateParameters)}
+     *
+     * @return a nullable {@link ExecutionStrategyInstrumentationContext} object that will be called back when the step ends (assuming it's not null)
+     */
+    @Nullable
+    default ExecuteObjectInstrumentationContext beginExecuteObject(InstrumentationExecutionStrategyParameters parameters, InstrumentationState state) {
+        return ExecuteObjectInstrumentationContext.NOOP;
     }
 
 
@@ -284,9 +300,25 @@ public interface Instrumentation {
      *
      * @return a nullable {@link InstrumentationContext} object that will be called back when the step ends (assuming it's not null)
      */
+    @Deprecated
+    @DeprecatedAt("2023-09-11")
     @Nullable
     default InstrumentationContext<ExecutionResult> beginField(InstrumentationFieldParameters parameters, InstrumentationState state) {
         return beginField(parameters.withNewState(state));
+    }
+
+    /**
+     * This is called just before a field is resolved into a value.
+     *
+     * @param parameters the parameters to this step
+     * @param state      the state created during the call to {@link #createState(InstrumentationCreateStateParameters)}
+     *
+     * @return a nullable {@link InstrumentationContext} object that will be called back when the step ends (assuming it's not null)
+     */
+    @Nullable
+    default InstrumentationContext<Object> beginFieldExecution(InstrumentationFieldParameters parameters, InstrumentationState state) {
+        InstrumentationContext<ExecutionResult> ic = beginField(parameters, state);
+        return ic == null ? null : new ExecutionResultInstrumentationContextAdapter(ic);
     }
 
     /**
@@ -343,9 +375,25 @@ public interface Instrumentation {
      *
      * @return a nullable {@link InstrumentationContext} object that will be called back when the step ends (assuming it's not null)
      */
+    @Deprecated
+    @DeprecatedAt("2023-09-11")
     @Nullable
     default InstrumentationContext<ExecutionResult> beginFieldComplete(InstrumentationFieldCompleteParameters parameters, InstrumentationState state) {
         return beginFieldComplete(parameters.withNewState(state));
+    }
+
+    /**
+     * This is called just before the complete field is started.
+     *
+     * @param parameters the parameters to this step
+     * @param state      the state created during the call to {@link #createState(InstrumentationCreateStateParameters)}
+     *
+     * @return a nullable {@link InstrumentationContext} object that will be called back when the step ends (assuming it's not null)
+     */
+    @Nullable
+    default InstrumentationContext<Object> beginFieldCompletion(InstrumentationFieldCompleteParameters parameters, InstrumentationState state) {
+        InstrumentationContext<ExecutionResult> ic = beginFieldComplete(parameters, state);
+        return ic == null ? null : new ExecutionResultInstrumentationContextAdapter(ic);
     }
 
     /**
@@ -372,9 +420,25 @@ public interface Instrumentation {
      *
      * @return a nullable {@link InstrumentationContext} object that will be called back when the step ends (assuming it's not null)
      */
+    @Deprecated
+    @DeprecatedAt("2023-09-11")
     @Nullable
     default InstrumentationContext<ExecutionResult> beginFieldListComplete(InstrumentationFieldCompleteParameters parameters, InstrumentationState state) {
         return beginFieldListComplete(parameters.withNewState(state));
+    }
+
+    /**
+     * This is called just before the complete field list is started.
+     *
+     * @param parameters the parameters to this step
+     * @param state      the state created during the call to {@link #createState(InstrumentationCreateStateParameters)}
+     *
+     * @return a nullable {@link InstrumentationContext} object that will be called back when the step ends (assuming it's not null)
+     */
+    @Nullable
+    default InstrumentationContext<Object> beginFieldListCompletion(InstrumentationFieldCompleteParameters parameters, InstrumentationState state) {
+        InstrumentationContext<ExecutionResult> ic = beginFieldListComplete(parameters, state);
+        return ic == null ? null : new ExecutionResultInstrumentationContextAdapter(ic);
     }
 
     /**