This is another PR on how ExecutionResult value wrapping could be avoided

See #3329  for a treatise on what's happening here


# Benchmarks

## master:
```
Benchmark                                    Mode  Cnt    Score   Error  Units
BenchMark.benchMarkSimpleQueriesThroughput  thrpt   15  181.263 ± 6.470  ops/s
BenchMark.benchMarkSimpleQueriesAvgTime      avgt   15    5.450 ± 0.226  ms/op
```

## This PR:
```
Benchmark                                    Mode  Cnt    Score   Error  Units
BenchMark.benchMarkSimpleQueriesThroughput  thrpt   15  185.092 ± 7.635  ops/s
BenchMark.benchMarkSimpleQueriesAvgTime      avgt   15    5.618 ± 0.367  ms/op
```

nominally better throughput but the time per op is no better for single threading benchmarks

I ran them with `@Thread(8)` on it and got the following


## master:
```
Benchmark                                    Mode  Cnt    Score     Error  Units
BenchMark.benchMarkSimpleQueriesThroughput  thrpt   15  802.114 ± 104.552  ops/s
BenchMark.benchMarkSimpleQueriesAvgTime      avgt   15    9.491 ±   0.536  ms/op
```

## This PR:
```

Benchmark                                    Mode  Cnt    Score    Error  Units
BenchMark.benchMarkSimpleQueriesThroughput  thrpt   15  817.685 ± 68.829  ops/s
BenchMark.benchMarkSimpleQueriesAvgTime      avgt   15    9.371 ±  0.304  ms/op
```

So faster and more throughput when the benchmarks run more threaded.

This PR would really be more about memory pressure created from the wrapping and unwrapping of ExecutionResults