@andimarek 2024-01-07
We should clearly call out the breaking changes for the release notes